reduce take/skip casts

+++ ExternDotnetSDK/ExternDotnetSDK/Client/Primitives/GetSliceAsync.cs
@@ -4,5 +4,5 @@ using System.Threading.Tasks;
 
 namespace Kontur.Extern.Client.Primitives
 {
-    internal delegate Task<(IReadOnlyList<T> Items, long totalItems)> GetSliceAsync<T>(long skip, long take, TimeSpan? timeout);
+    internal delegate Task<(IReadOnlyList<T> Items, long totalItems)> GetSliceAsync<T>(long skip, int take, TimeSpan? timeout);
 }
